Changed the create team page into a dialog window and removed the old way of creating teams.
