+++ src/App.vue
@@ -87,7 +87,7 @@
         <v-list>
           <v-list-tile v-if="isConfigured" class="mt-3">
             <v-list-tile-action>
-              <v-btn icon ripple @click="createTeam">
+              <v-btn icon ripple @click="createTeamDialog = true">
                 <v-icon color="grey darken-1">add_circle_outline</v-icon>
               </v-btn>
             </v-list-tile-action>
@@ -148,6 +148,19 @@
         </v-container>
       </v-content>
       <snackbar/>
+      <v-dialog v-model="createTeamDialog" width="500">
+          <v-card>
+            <v-toolbar dark>
+                <v-btn icon dark @click="createTeamDialog = false">
+                <v-icon>close</v-icon>
+                </v-btn>
+                <v-toolbar-title class="white--text">Create a new team</v-toolbar-title>
+            </v-toolbar>
+            <v-card-text>
+              <CreateTeam />
+            </v-card-text>
+          </v-card>
+      </v-dialog>
     </v-app>
   </div>
 </template>
@@ -156,17 +169,16 @@ import _ from "lodash";
 import { mapGetters } from "vuex";
 import Snackbar from "@/components/ui/Snackbar.vue";
 import TeamAvatar from "@/components/TeamAvatar.vue";
+import CreateTeam from "@/components/CreateTeam.vue";
 
 export default {
   data: () => ({
-    drawer: null
+    drawer: null,
+    createTeamDialog: false
   }),
   methods: {
     openSettings() {
       this.$router.push({ name: "settings" });
-    },
-    createTeam() {
-      this.$router.push({ name: "create-team" });
     },
     deleteTeam(team) {
       this.$store.dispatch("deleteTeam", team);
@@ -210,7 +222,8 @@ export default {
   },
   components: {
     Snackbar,
-    TeamAvatar
+    TeamAvatar,
+    CreateTeam
   }
 };
 </script>
